% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_master_list.R
\name{get_master_list}
\alias{get_master_list}
\title{Function to get master list of drugs from LiverTox}
\usage{
get_master_list()
}
\value{
A data frame containing the master list of drugs from LiverTox
}
\description{
Function to get master list of drugs from LiverTox
}
\examples{
df <- get_master_list()
}
